backend/users/requirements.txt
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.1
billiard==4.2.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Deprecated==1.2.14
ecdsa==0.18.0
fastapi==0.108.0
flake8==7.1.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
pytest==8.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.32.0.post1
tornado==6.4
typing_extensions==4.12.2
uvicorn==0.25.0
uvloop==0.19.0
wcwidth==0.2.13
wrapt==1.16.0
